requirements.txt
astroid==2.3.0
aws-cfn-bootstrap==2.0
backcall==0.2.0
botocore==1.21.54
decorator==5.1.0
Django==2.0.2
docutils==0.14
git-remote-codecommit==1.15.1
ikp3db==1.4.1
importlib-metadata==4.8.1
ipython==7.28.0
isort==4.3.21
jedi==0.18.0
jmespath==0.10.0
lazy-object-proxy==1.6.0
lockfile==0.11.0
matplotlib-inline==0.1.3
mccabe==0.6.1
parso==0.8.2
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
ptyprocess==0.7.0
Pygments==2.10.0
pylint==2.4.4
pylint-django==2.3.0
pylint-flask==0.6
pylint-plugin-utils==0.6
pystache==0.5.4
python-daemon==2.2.3
python-dateutil==2.8.2
pytz==2021.3
simplejson==3.2.0
six==1.16.0
stevedore==3.4.0
traitlets==5.1.0
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.7
virtualenv==16.2.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.5
wrapt==1.13.1
zipp==3.6.0
